Revert "Beans"
