+++ nbproject/project.properties
@@ -3,7 +3,6 @@ annotation.processing.enabled.in.editor=true
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-auxiliary.org-netbeans-modules-projectapi.jsf_2e_language=Facelets
 build.classes.dir=${build.web.dir}/WEB-INF/classes
 build.classes.excludes=**/*.java,**/*.form
 build.dir=build
@@ -38,8 +37,7 @@ j2ee.platform.classpath=${j2ee.server.home}/lib/annotations-api.jar:${j2ee.serve
 j2ee.server.type=Tomcat
 jar.compress=false
 javac.classpath=\
-    ${libs.JDBC_PostgreSQL.classpath}:\
-    ${libs.jsf20.classpath}
+    ${libs.JDBC_PostgreSQL.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.debug=true
