--- src/java/beans/ServLogin.java
@@ -1,103 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package beans;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-import static java.lang.System.out;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.swing.JOptionPane;
-import pq1.OperUser;
-import pq3.Usuario;
-
-/**
- *
- * @author santiagob20
- */
-@WebServlet(name = "/login", urlPatterns = {"/Login"})
-public class ServLogin extends HttpServlet {
-
-    /**
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-     * methods.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    
-    Usuario u = new Usuario();
-   
-    
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        OperUser op = new OperUser();
-       
-        response.setContentType("text/html;charset=UTF-8");
-      
-       u.setUser(request.getParameter("txtNombre"));
-       u.setPassword(request.getParameter("txtContrasena"));
-       boolean rta = op.consultar(u);
-       
-       if(rta)
-        {
-            response.sendRedirect("Principal.html");
-        }
-         else
-         {
-             response.sendRedirect("index.html");
-             //JOptionPane.showMessageDialog(null, "¡Usuario o contraseña invalidos!",null,JOptionPane.ERROR_MESSAGE);
-         }
-        
-    }
-
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /**
-     * Handles the HTTP <code>GET</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        processRequest(request, response);
-    }
-
-    /**
-     * Handles the HTTP <code>POST</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        processRequest(request, response);
-    }
-
-    /**
-     * Returns a short description of the servlet.
-     *
-     * @return a String containing servlet description
-     */
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }// </editor-fold>
-
-}
-    
-
